First crack at transit info
new.mta.info and subwaytime.mta.info are two online providers of MTA
service times at transit stations. The former is a OpenTripPlanner
instance that subscribes to all relevant agencies in the metro area
and includes GTFS-RT in addition to plain GTFS schedules. The latter
appears to be a proprietary node.js application (per exceptions it
generates) that seems to return real time subway status only.

The OTP instance configuration appears to have a bug at several
terminus stations where it only shows arriving trains but not
departing trains. This is a weird bug because planning routes with the
/plan API includes up-to-date departure data.  The proprietary node.js
app has no such issues, but its API is far less rich or
standardized. So this commit populates a :transit key and a
:transit-fallback key corresponding to these two sources,
respectively. Views can be wired to either depending on the data
quality. Ideally, the MTA will fix the issue with terminus stations
and the fallback can be dropped.

This needs several improvements:

- Ideally the subscriptions should return more than just the remaining
  minutes
- It currently assumes one route per station, which is not correct for
  most cases
- Since the fallback includes both directions and the OTP stoptimes
  prequest includes only one direction (and directions are
  disambiguated with different strings), there's some awkwardness is
  handling this mismatch